Return a non-zero exit code when copying fails
Even when the requested copy operation fails, the copy_file() function
returns the value returned by the mtd_unmount() function, which is
almost guaranteed to succeed.  Since the value returned by the
copy_file() function is carried over to the program's exit code, this
logic results in returning an exit code indicating success even when the
program fails to correctly carry out the requested copy operation.

Tweak the copy_file() function so that it returns a non-zero value when
the requested copy operation fails.  Change the return type of the
mtd_unmount() function to void as there is pretty much nothing that can
be done to fix any potential problems it reports (log messages will
still be generated in case of problems).