Use inband tags when required
If the MTD provided has less bytes available in its OOB area than are
necessary to store a full Yaffs2 tag structure (including tags ECC data)
and Yaffs2 is used, inband tags must be enabled.  This makes Yaffs2 code
expect tags to be present in the flash data area, alongside the data,
rather than in the OOB area.

Note that even if there are enough bytes available in the OOB area to
store a full Yaffs2 tag structure, inband tags may be used anyway, but
this change does not handle that case.