Fetch the number of available OOB bytes from sysfs
In order to detect whether use of inband tags is required, the number of
available OOB bytes needs to be determined for the MTD provided.  That
parameter is not returned by the MEMGETINFO ioctl, so it has to be
retrieved from sysfs.

Add code that reads the contents of /sys/class/mtd/mtd<X>/oobavail,
converts it to a number, and passes the latter to the init_yaffs_dev()
function, where it will be used for determining whether use of inband
tags is required.