Enable NOR flash autodetection
NOR flash does not handle MEMREAD/MEMWRITE ioctls because it does not
have a spare area.  For that reason, the Yaffs driver must be able to
differentiate between NAND flash and NOR flash so that appropriate
helper functions are used for interacting with each of these device
classes.

Add a new 'mtd_type' member to the structure passed around by the Yaffs
driver.  Set it to the detected flash memory type in src/mtd.c,
specifically in init_yaffs_geometry(), which now also overrides the
chunk size and block size autodetected for NOR flash, so that this MTD
type can be used by Yaffs code.  Since NOR flash does not have a spare
area and is byte-addressable, the Yaffs layout used on such hardware can
be fairly arbitrary; the defaults assumed are those used by Yaffs code
when creating Yaffs2 file system images, but they can be overridden by
the user via command-line options -C and -B.