Make data offset calculation more flexible
Implement a new helper function, ydrv_get_data_offset_for_chunk(), which
calculates the MTD offset at which data starts for a specific Yaffs
chunk.  While more complicated than a simple multiplication, the new
function handles Yaffs layouts in which the block size is not a multiple
of the chunk size (with padding between the last chunk in a block and
the first chunk of the following block).  While such layouts do not
really make sense for NAND flash, they can be used on NOR flash.  Use
the new function where applicable to make offset calculations
consistent.