Support disabling ECC for tags
Some Yaffs2 file systems do not use ECC for tags.  Yafut needs to handle
such file systems.

Add a new command-line option, -E, which allows use of ECC for tags to
be disabled.  Ensure the logic used for autodetecting use of inband tags
also accounts for the value of this new option.

Update README.md to explain in what scenario the new option
may be useful.