Extend the scope of init_yaffs_geometry()
Pull the discover_mtd_parameters() call into init_yaffs_geometry() so
that the latter encompasses the entire logic related to establishing the
Yaffs geometry to use.  In preparation for adding support for operating
on regular files, move most of the logic previously contained in that
function to init_yaffs_geometry_nand_or_nor(), to emphasize that the
latter only applies to MTD character devices.