Support forcing byte order
Add two new command-line options, -L and -M, which allow forcing Yaffs
code to assume little-endian or big-endian byte order, respectively.
This enables little-endian hosts to operate on big-endian file system
and vice versa.  If neither of the new options is used, host CPU byte
order is assumed by default.  Update README.md accordingly.

Note that Yaffs2 source code must be patched in order for cross-endian
operations to work for file systems that include ECC data for tags.
Unfortunately, Yaffs1 code is affected by the same issue in a different
location, where it is not as easy to fix as for Yaffs2, nor is it likely
that it ever will be as Yaffs1 code is currently pretty much set in
stone.  Given that, a Yaffs1 patch for this deficiency was never
prepared and Yafut is currently only able to operate on Yaffs2 file
systems in an endian-portable manner.