Terminate storage driver array with a canary value
Since the list of storage drivers included in the build needs to be made
platform-specific, the array holding pointers to supported storage
drivers will need to be moved to a separate code module, which will in
turn break the currently implemented storage driver matching logic (as
the size of the array will not be known at compile time).  Use a canary
value (NULL) at the end of that array to enable storage driver matching
to also work when the variables it uses are split across multiple code
modules.