Hide struct mtd_info_user behind an opaque pointer
MTD support in Yafut is Linux-specific.  To enable the tool to be built
on other operating systems, platform-specific parts of the source code
must be contained in conditionally-compiled modules.

In preparation for that, replace struct mtd_info_user in struct
storage_probe_info with an opaque pointer to a dynamically allocated
chunk of memory.  Adjust Linux-specific storage drivers (for NAND and
NOR MTDs) to use that pointer accordingly.