Support preserving file modification times
When the Yaffs Direct Interface is used, replacing the contents of an
existing file on a Yaffs file system does not automatically cause its
last modification time to be updated.

Add a new command-line option, -t, which allows preserving file
modification times when copying files from/to a Yaffs file system.  This
new feature is not enabled by default because updating a file
modification timestamp on a Yaffs file system requires writing an extra
data chunk to the underlying storage device.