Help static analyzers understand Yaffs error code
Clang Static Analyzer is unable to infer that yaffsfs_GetLastError() is
guaranteed to return a negative integer, which causes the tool to report
false positives about variables potentially being uninitialized in some
scenarios.  Add a wrapper for yaffsfs_GetLastError() that is similar in
spirit to util_get_errno(), but operates on the error code set by Yaffs.